Update patch set 3
Patch Set 3: Code-Review-1

> Patch Set 3:
> 
> @Patrick
> Do you have any comments?

Yes. You still haven’t resolved my comment. 

This should not be done in a bbappend but in a machine.conf file. Doing it in a bbappend means it only applies for this recipe but other recipes use the variable. It needs to be applied everywhere.

Patch-set: 3
Reviewer: Gerrit User 1000000 <1000000@adbd0a64-1f21-4d83-b585-671fe73cb6e4>
Label: Code-Review=-1